+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 * Version 1.3.4:
     * Set recursion depth to 5000
     * Remove internal labels in pruned trees
-    * Allow output prefix
+    * Allow output prefix and allow outputing to an existing folder
     * Add option to show version
     * Show help message if no argument is provided
+    * Prevent conversion to scientific form
 * Version 1.3.3:
 	* Suppress quotes around dashes in the output trees
 * Version 1.3.2: